Only load `NavMobileClient` if user is on mobile
